Moving custom exceptions to a file with their own classes
