+++ lib/rtoolsHCK.rb
@@ -1,6 +1,7 @@
 #!/usr/bin/env ruby
 # frozen_string_literal: true
 
+require_relative 'exceptions'
 require 'net/telnet'
 require 'securerandom'
 require 'winrm'
@@ -68,26 +69,6 @@ class RToolsHCK
   WINRM_RECIEVE_TIMEOUT = 99_999
 
   private
-
-  # A custom RToolsHCK error exception
-  class RToolsHCKError < StandardError
-    # Custom addition to the exception backtrace, (better logging)
-    attr_reader :where
-
-    # Initialization of the custom exception
-    def initialize(where)
-      @where = where
-    end
-  end
-
-  # A custom RToolsHCK connection error exception
-  class RToolsHCKConnectionError < RToolsHCKError; end
-
-  # A custom RToolsHCK action error exception
-  class RToolsHCKActionError < RToolsHCKError; end
-
-  # A custom Winrm powershell run error exception
-  class WinrmPSRunError < RToolsHCKActionError; end
 
   def handle_exceptions
     yield
